+++ mooc-authorize2/src/main/java/cn/edu/lingnan/authorize/dao/UserDAO.java
@@ -106,7 +106,7 @@ public class UserDAO extends BaseDAO {
                 user.getPassword(),user.getStatus());
     }
 
-    public void updateLoginTime(Integer id, Date loginTime){
+    public void updateLoginTime(Long id, Date loginTime){
         String sql = "update mooc_user set login_time = ? where id = ?";
         jdbcTemplate.update(sql,loginTime,id);
     }
